docs(tutorial): move login to pages
