+++ documentation/tutorial/next-steps/inferencer/react-router/ant-design/index.md
@@ -80,7 +80,7 @@ import { CreateProduct } from "./pages/products/create";
 // highlight-next-line
 import { ListCategories } from "./pages/categories/list";
 
-import { Login } from "./components/login";
+import { Login } from "./pages/login";
 
 import "antd/dist/reset.css";
