+++ documentation/tutorial/routing/intro/react-router/index.md
@@ -56,7 +56,7 @@ import { EditProduct } from "./pages/products/edit";
 import { ListProducts } from "./pages/products/list";
 import { CreateProduct } from "./pages/products/create";
 
-import { Login } from "./components/login";
+import { Login } from "./pages/login";
 import { Header } from "./components/header";
 
 export default function App(): JSX.Element {
