+++ documentation/tutorial/routing/resource-definition/react-router/index.md
@@ -44,7 +44,7 @@ import { EditProduct } from "./pages/products/edit";
 import { ListProducts } from "./pages/products/list";
 import { CreateProduct } from "./pages/products/create";
 
-import { Login } from "./components/login";
+import { Login } from "./pages/login";
 import { Header } from "./components/header";
 
 export default function App(): JSX.Element {
@@ -133,7 +133,7 @@ import { EditProduct } from "./pages/products/edit";
 import { ListProducts } from "./pages/products/list";
 import { CreateProduct } from "./pages/products/create";
 
-import { Login } from "./components/login";
+import { Login } from "./pages/login";
 import { Header } from "./components/header";
 
 export default function App(): JSX.Element {
