+++ documentation/tutorial/routing/resource-definition/react-router/sandpack.tsx
@@ -36,7 +36,7 @@ import { EditProduct } from "./pages/products/edit";
 import { ListProducts } from "./pages/products/list";
 import { CreateProduct } from "./pages/products/create";
 
-import { Login } from "./components/login";
+import { Login } from "./pages/login";
 import { Header } from "./components/header";
 
 export default function App(): JSX.Element {
@@ -97,7 +97,7 @@ import { EditProduct } from "./pages/products/edit";
 import { ListProducts } from "./pages/products/list";
 import { CreateProduct } from "./pages/products/create";
 
-import { Login } from "./components/login";
+import { Login } from "./pages/login";
 import { Header } from "./components/header";
 
 export default function App(): JSX.Element {
