+++ documentation/tutorial/ui-libraries/intro/ant-design/react-router/sandpack.tsx
@@ -39,7 +39,7 @@ import { EditProduct } from "./pages/products/edit";
 import { ListProducts } from "./pages/products/list";
 import { CreateProduct } from "./pages/products/create";
 
-import { Login } from "./components/login";
+import { Login } from "./pages/login";
 import { Header } from "./components/header";
 
 // We're importing a reset.css file to reset the default styles of the browser.
@@ -124,7 +124,7 @@ import { EditProduct } from "./pages/products/edit";
 import { ListProducts } from "./pages/products/list";
 import { CreateProduct } from "./pages/products/create";
 
-import { Login } from "./components/login";
+import { Login } from "./pages/login";
 
 // We're importing a reset.css file to reset the default styles of the browser.
 import "antd/dist/reset.css";
